Add GNU Readline functionality to command prompt
This commit brings GEMPAK into the 1990's by calling on the GNU
Readline library to handle the GEMPAK prompt.  This allows for the
arrow keys to work as you would expect, as well as other keyboard
shortcuts.  We assume the Readline library exists in a standard
directory.  Only Linux is supported with the edits here.